Rename utility to `isValueUsed`
