+++ rules/prefer-node-remove.js
@@ -1,6 +1,6 @@
 'use strict';
 const getDocumentationUrl = require('./utils/get-documentation-url');
-const isReturnValueUsed = require('./utils/is-return-value-used');
+const isValueUsed = require('./utils/is-value-used');
 
 const getMethodName = callee => {
 	const {property} = callee;
@@ -67,7 +67,7 @@ const create = context => {
 
 				let fix = fixer => fixer.replaceText(node, `${argumentName}.remove()`);
 
-				if (!isReturnValueUsed(node)) {
+				if (!isValueUsed(node)) {
 					fix = undefined;
 				}
